+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+## [1.5.0] - 2021-04-15
+
 ### Added
 
 - Support for Furucombo rCOMBO.